Reset parsed content scripts

+++ app/src/com/constellio/app/extensions/api/scripts/ResetParsedContentScript.java
@@ -3,14 +3,11 @@ package com.constellio.app.extensions.api.scripts;
 import com.constellio.app.modules.rm.services.RMSchemasRecordsServices;
 import com.constellio.app.modules.rm.wrappers.Document;
 import com.constellio.app.services.factories.AppLayerFactory;
-import com.constellio.app.services.migrations.ConstellioEIM;
 import com.constellio.data.dao.services.bigVault.SearchResponseIterator;
 import com.constellio.data.io.services.facades.IOServices;
 import com.constellio.model.entities.records.Record;
-import com.constellio.model.entities.records.wrappers.UserDocument;
 import com.constellio.model.entities.schemas.Schemas;
 import com.constellio.model.services.contents.ContentManager;
-import com.constellio.model.services.records.SchemasRecordsServices;
 import com.constellio.model.services.search.SearchServices;
 import com.constellio.model.services.search.query.logical.LogicalSearchQuery;
 import com.constellio.model.services.search.query.logical.LogicalSearchQueryOperators;
@@ -35,15 +32,15 @@ public class ResetParsedContentScript extends ScriptWithLogOutput {
         for(String collection : appLayerFactory
                 .getCollectionsManager().getCollectionCodesExcludingSystem())
         {
-            RMSchemasRecordsServices schemasRecordsServices = new RMSchemasRecordsServices(collection, modelLayerFactory);
+            RMSchemasRecordsServices rmSchemasRecordsServices = new RMSchemasRecordsServices(collection, appLayerFactory);
             LogicalSearchQuery logicalSearchQuery = new LogicalSearchQuery();
-            logicalSearchQuery.setCondition(LogicalSearchQueryOperators.from(schemasRecordsServices.userDocument.schemaType()).returnAll());
+            logicalSearchQuery.setCondition(LogicalSearchQueryOperators.from(rmSchemasRecordsServices.documentSchemaType()).returnAll());
             SearchResponseIterator<Record> searchResponseIterator = searchServices.recordsIterator(logicalSearchQuery, 2000);
             ContentManager contentManager = modelLayerFactory.getContentManager();
 
             while(searchResponseIterator.hasNext()) {
                  Record currentRecord = searchResponseIterator.next();
-                 Document document = schemasRecordsServices.wrapDocument(currentRecord);
+                 Document document = rmSchemasRecordsServices.wrapDocument(currentRecord);
 
                  String fileName = contentManager.getParsedContentFileName(document.getContent().getCurrentVersion().getHash());
                  File parsedFilePath = contentManager.getContentDao().getFileOf(fileName);
@@ -56,8 +53,13 @@ public class ResetParsedContentScript extends ScriptWithLogOutput {
             }
 
             for(File currentFileToSuppress : documentToSuppress) {
-                ioServices.deleteQuietly(currentFileToSuppress);
+                if (currentFileToSuppress.exists()) {
+                    ioServices.deleteQuietly(currentFileToSuppress);
+                    outputLogger.appendToFile("Deleted file : " + currentFileToSuppress.getName() + "\n");
+                }
             }
         }
+
+        outputLogger.appendToFile("Execution done. Deleted : " + documentToSuppress.size() + " parsed document");
     }
 }
